[Add] Новый тип данных, копирование полей, задание сетевого адреса [Fix] Дробное значение делителя в конфигурации

+++ Drivers/MODBUS_TCP/Models/PackagePart.cs
@@ -6,7 +6,7 @@
 
         public string Type { get; set; } = "";
 
-        public ushort Scale { get; set; } = 1;
+        public float Scale { get; set; } = 1;
 
         public byte Length { get; set; } = 0;
